Moved reusable validators into generic concern
