+++ app/models/cash_contribution.rb
@@ -1,4 +1,6 @@
 class CashContribution < ApplicationRecord
+  include GenericValidator
+
   self.implicit_order_column = "created_at"
   belongs_to :project
   has_many_attached :cash_contribution_evidence_files
@@ -19,24 +21,6 @@ class CashContribution < ApplicationRecord
         50,
         "Description of your cash contribution must be 50 words or fewer"
     )
-
-  end
-
-  def validate_file_attached(field, error_msg)
-    unless self.public_send(field).attached?
-      errors.add(field, error_msg)
-    end
-  end
-
-  def validate_length(field, max_length, error_msg)
-
-    word_count = self.public_send(field)&.split(' ')&.count
-
-    logger.debug "#{field} word count is #{word_count}"
-
-    if word_count && word_count > max_length
-      self.errors.add(field, error_msg)
-    end
 
   end
