+++ app/models/project.rb
@@ -1,5 +1,5 @@
 class Project < ApplicationRecord
-    include ActiveModel::Validations
+    include ActiveModel::Validations, GenericValidator
     self.implicit_order_column = "created_at"
 
     belongs_to :user
@@ -274,24 +274,6 @@ class Project < ApplicationRecord
         no: 1,
         x_not_sure: 2
     }
-
-    def validate_file_attached(field, error_msg)
-        unless self.public_send(field).attached?
-            errors.add(field, error_msg)
-        end
-    end
-
-    def validate_length(field, max_length, error_msg)
-
-        word_count = self.public_send(field)&.split(' ')&.count
-
-        logger.debug "#{field} word count is #{word_count}"
-
-        if word_count && word_count > max_length
-            self.errors.add(field, error_msg)
-        end
-
-    end
 
     def to_salesforce_json
         Jbuilder.encode do |json|
